+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+* Add #started_at, #ends_at, and #seconds_remaining to Song
 * Add dj_escorted_off / dj_booed_off events
 * Add fan_added / fan_removed events
 * Add user_name_updated / user_avatar_updated events